DEMO_CONTAINER/scripts/requirements.txt
google-auth==1.14.0
google-auth-oauthlib==0.4.1
google-cloud-core==1.3.0
google-cloud-storage==1.27.0
pandas